==> ml/notebooks/02-model-training.ipynb <==
# imports 

! pip install xgboost
from pathlib import Path
import xgboost as xgb
import pandas as pd
import numpy as np
import sagemaker
import botocore
import joblib
import boto3
import yaml
import json
import sys
import os
from time import gmtime, strftime, sleep
from sagemaker.experiments.run import Run, load_run

sagemaker.__version__

boto_session = boto3.Session()
region = boto_session.region_name
print(f"Region--{region}")
sm_role = sagemaker.get_execution_role()
print(f"Sagemaker Role--{sm_role}")
sagemaker_session = sagemaker.Session()
default_bucket = sagemaker_session.default_bucket()
print(f"Default bucket--{default_bucket}")

# get data

import io
s3_client = boto3.client("s3")
file_key  = 'bits-webminar-june-24/bike-sharing-dataset.csv'
print(file_key)
response = s3_client.get_object(Bucket=default_bucket, Key=file_key)
file_content = response['Body'].read()
df = pd.read_csv(io.BytesIO(file_content), encoding='utf-8')
df.head(3)

import sys
root_dir = Path('../scripts').resolve().parents[0] 
print(root_dir)
sys.path.append(str(root_dir))


# feature engineering

import scripts.features as features
processed_data = features.pre_process(df)


processed_data

# create a sagemaker experiment 

from datetime import datetime, timedelta
def utc_to_ist(utc_dt):
    return utc_dt + timedelta(hours=5, minutes=30)

def get_current_ist():
    current_utc_time = datetime.utcnow()
    current_ist_time = utc_to_ist(current_utc_time)
    return current_ist_time

# current_ist_time = get_current_ist()
# experiment_name = current_ist_time.strftime("bits-webminar-june-24-%d-%m-%Y-%H-%M-%S")
experiment_name = current_ist_time.strftime("bike-ride-sharing")
experiment_name

# split data --once

# Shuffle and splitting dataset
train_data, validation_data, test_data = np.split(
    processed_data.sample(frac=1, random_state=1729),
    [int(0.7 * len(processed_data)), int(0.9 * len(processed_data))],
)

print(f"Data split > train:{train_data.shape} | validation:{validation_data.shape} | test:{test_data.shape}")

def upload_file_to_s3(df,file_key):
    # Convert DF to csv buffer
    csv_buffer = io.StringIO()
    df.to_csv(csv_buffer)
    BUCKET_NAME  = default_bucket
    PREFIX = 'bits-webminar-june-24/'
    KEY = f"{PREFIX}{file_key}"
    try:
        s3_client.put_object(
            Body=csv_buffer.getvalue(), 
            Bucket = BUCKET_NAME, 
            Key = KEY,
        )
        print("Upload Successful")
    except FileNotFoundError:  
        print("The file was not found")
    except botocore.exceptions.ClientError as error:
        if error.response['Error']['Code'] == "404":
            print("The object does not exist")
        else:
            raise

upload_file_to_s3(train_data,'train_data.csv')
upload_file_to_s3(validation_data,'validation_data.csv')
upload_file_to_s3(test_data,'test_data.csv')

# get data from s3 

def get_data_from_s3_as_df(file_key):
    BUCKET_NAME  = default_bucket
    PREFIX = 'bits-webminar-june-24/'
    KEY = f'{PREFIX}{file_key}'
    print(f"downloadig..{KEY}")
    response = s3_client.get_object(Bucket=BUCKET_NAME, Key=KEY)
    file_content = response['Body'].read()
    df = pd.read_csv(io.BytesIO(file_content), encoding='utf-8')
    return df 

train_data = get_data_from_s3_as_df('train_data.csv')
test_data  = get_data_from_s3_as_df('test_data.csv')
validation_data = get_data_from_s3_as_df('validation_data.csv')

train_data.drop(['Unnamed: 0'],axis=1 , inplace=True)
test_data.drop(['Unnamed: 0'],axis=1 , inplace=True)
validation_data.drop(['Unnamed: 0'],axis=1 , inplace=True)

# model training and validation

# target_col = config['target']
target_col = 'cnt'
target_col

train_features = train_data.drop(target_col, axis=1)
train_label = pd.DataFrame(train_data[target_col])

dtrain = xgb.DMatrix(train_features, label=train_label)

# Set hyperparameters for regression
hyperparams = {
    "max_depth": 5,
    "eta": 0.5,
    "alpha": 2.5,
    "objective": "reg:squarederror",  # objective regression
    "subsample": 0.8,
    "colsample_bytree": 0.8,
    "min_child_weight": 3
}

num_boost_round = 150
nfold = 3
early_stopping_rounds = 10

# Cross-validate on training data
cv_results = xgb.cv(
    params=hyperparams,
    dtrain=dtrain,
    num_boost_round=num_boost_round,
    nfold=nfold,
    early_stopping_rounds=early_stopping_rounds,
    metrics=["rmse"],  #  regression metrics RMSE
    seed=10,
)

cv_results

# metrics data
metrics_data = {
    "regression_metrics": {
        "validation:rmse": {
            "value": cv_results.iloc[-1]["test-rmse-mean"],
            "standard_deviation": cv_results.iloc[-1]["test-rmse-std"]
        },
        "train:rmse": {
"value": cv_results.iloc[-1]["train-rmse-mean"],
"standard_deviation": cv_results.iloc[-1]["train-rmse-std"]
        },
    }
}

print(f"Cross-validated train-rmse:{cv_results.iloc[-1]['train-rmse-mean']:.2f}")
print(f"Cross-validated validation-rmse:{cv_results.iloc[-1]['test-rmse-mean']:.2f}")

# training model on whole dataset

test_features = test_data.drop(target_col, axis=1)
test_label = pd.DataFrame(test_data[target_col])
dtest = xgb.DMatrix(test_features, label=test_label)

# create a Run

current_ist_time = get_current_ist()
current_ist_time

run_suffix = current_ist_time.strftime("%d-%m-%Y-%H-%M-%S")
run_suffix

sagemaker_session = sagemaker.Session()
sm_client = sagemaker_session.sagemaker_client

# get file paths from s3 to log in the run 

objects = s3_client.list_objects_v2(Bucket=default_bucket)
for obj in objects.get('Contents', []):
    # print(obj)
    file_key = obj['Key']
    # print(file_key)
    if file_key == "bits-webminar-june-24/":
        continue
    else:
        s3_file_path = f"s3://{default_bucket}/{file_key}"
        print(s3_file_path)



with Run(experiment_name=experiment_name,
         run_name=f"training-{run_suffix}",
         run_display_name="training",
         sagemaker_session=sagemaker_session) as run:
    run.log_parameters(
        {
            "train": 0.7,
            "validate": 0.2,
            "test": 0.1
        }
    )
    # Log input dataset metadata and output
    # Log files from S3
    run.log_artifact(name='bike-share-dataset',value='s3://sagemaker-eu-north-1-058264393695/bike-sharing-dataset.csv', media_type="text/csv")
    run.log_artifact(name= 'train-csv',value='s3://sagemaker-eu-north-1-058264393695/bits-webminar-june-24/train_data.csv', media_type="text/csv")
    run.log_artifact(name = 'test-csv',value='s3://sagemaker-eu-north-1-058264393695/bits-webminar-june-24/test_data.csv', media_type="text/csv")
    run.log_artifact(name = 'validation-csv',value='s3://sagemaker-eu-north-1-058264393695/bits-webminar-june-24/validation_data.csv', media_type="text/plain")


from sklearn.metrics import mean_squared_error

print(f"run_suffix--{run_suffix}")

# Train the model for different max_depth values
for i, d in enumerate([2, 5, 10, 15, 20]):
    hyperparams["max_depth"] = d
    
    print(f"Fit estimator with max_depth={d}")
    run_name = f"training-{i}-{run_suffix}"
    
    with Run(experiment_name=experiment_name,
             run_name=run_name,
             run_display_name=f"max-depth-{d}",
             sagemaker_session=sagemaker_session) as run:
        # Train the model
        model = (
            xgb.train(
                params=hyperparams, 
                dtrain=dtrain, 
                evals = [(dtrain,'train'), (dtest,'eval')], 
                num_boost_round=num_boost_round, 
                early_stopping_rounds=early_stopping_rounds, 
                verbose_eval = 0
            )
        )
        # Calculate regression metrics 
        test_rmse = np.sqrt(mean_squared_error(test_label, model.predict(dtest)))
        train_rmse = np.sqrt(mean_squared_error(train_label, model.predict(dtrain)))
        
        # Log metrics to the run
        run.log_parameters(hyperparams)
        run.log_metric(name="test_rmse", value = test_rmse, step=d)
        run.log_metric(name="train_rmse", value = train_rmse, step=d)

        # time.sleep(8) # wait until resource tags are propagated to the run
        print(f"Test RMSE: {test_rmse:.4f} | Train RMSE: {train_rmse:.4f}")
        

# delete a sagemaker experiment 

import boto3
sagemaker = boto3.client('sagemaker')



def disassociate_trial_components(trial_name):
    # List trial components associated with the trial
    response = sagemaker.list_trial_components(TrialName=trial_name)
    trial_components = response['TrialComponentSummaries']

    # Disassociate each trial component from the trial
    for component in trial_components:
        component_arn = component['TrialComponentArn']
        print(f"Disassociating trial component: {component_arn} from trial: {trial_name}")
        sagemaker.disassociate_trial_component(
            TrialComponentName=component['TrialComponentName'],
            TrialName=trial_name
        )

def delete_trial_components(trial_name):
    # List trial components associated with the trial
    response = sagemaker.list_trial_components(TrialName=trial_name)
    trial_components = response['TrialComponentSummaries']

    # Delete each trial component
    for component in trial_components:
        component_name = component['TrialComponentName']
        print(f"Deleting trial component: {component_name}")
        sagemaker.delete_trial_component(TrialComponentName=component_name)

def delete_trials(experiment_name):
    # List trials in the experiment
    trials_response = sagemaker.list_trials(ExperimentName=experiment_name)
    trials = trials_response['TrialSummaries']

    # Delete each trial
    for trial in trials:
        trial_name = trial['TrialName']
        print(f"Disassociating trial components for trial: {trial_name}")
        disassociate_trial_components(trial_name)
        print(f"Deleting trial components for trial: {trial_name}")
        delete_trial_components(trial_name)
        print(f"Deleting trial: {trial_name}")
        sagemaker.delete_trial(TrialName=trial_name)

def delete_experiment(experiment_name):
    try:
        print(f"Deleting trials for experiment: {experiment_name}")
        delete_trials(experiment_name)
        
        # Delete the experiment
        print(f"Deleting experiment: {experiment_name}")
        sagemaker.delete_experiment(ExperimentName=experiment_name)
        print(f"Experiment {experiment_name} deleted successfully.")
    except sagemaker.exceptions.ResourceNotFound as e:
        print(f"Error: {e}")
    except Exception as e:
        print(f"Unexpected error: {e}")

# Replace 'your-experiment-name' with the name of your experiment
# experiment_name = 'bits-build-06-21-11-37-28-p-ljnvfgbcukgr'
experiment_name = 'bits-poc-pipeline'


delete_experiment(experiment_name)


# check job status

import boto3
sagemaker_client = boto3.client('sagemaker')
response = sagemaker_client.list_processing_jobs()
processing_job_names = []
for job in response['ProcessingJobSummaries']:
    print(job['ProcessingJobName'])
    processing_job_names.append(job['ProcessingJobName'])
# processing_job_names

# processing_job_names

# processing_job_names = [
# 'container-processing-2024-06-22-05-19-50-659',
# 'pipelines-y4ww6ymvc9rf-bits-webminar-june-2-aKUFROKNFA',
# 'container-processing-2024-06-21-11-24-07-988',
# 'container-processing-2024-06-21-11-08-38-645',
# 'container-processing-2024-06-21-10-53-06-841',
# 'container-processing-2024-06-21-10-35-38-917',
# 'container-processing-2024-06-21-09-58-42-857'
# ]


def stop_processing_jobs(job_names):
    for job_name in job_names:
        try:
            # Stop the processing job if it is still running
            sagemaker_client.stop_processing_job(ProcessingJobName=job_name)
            print(f"Stopped processing job: {job_name}")
        except sagemaker_client.exceptions.ClientError as e:
            if 'ValidationException' in str(e):
                print(f"Job {job_name} is already stopped or does not exist.")
            else:
                raise

stop_processing_jobs(processing_job_names)

def check_processing_job_status(job_names):
    for job_name in job_names:
        try:
            response = sagemaker_client.describe_processing_job(ProcessingJobName=job_name)
            status = response['ProcessingJobStatus']
            print(f"Job {job_name} status: {status}")
        except sagemaker_client.exceptions.ClientError as e:
            if 'ValidationException' in str(e):
                print(f"Job {job_name} does not exist.")
            else:
                raise

check_processing_job_status(processing_job_names)


# def delete_processing_jobs(job_names):
#     for job_name in job_names:
#         try:
#             sagemaker_client.stop_processing_job(ProcessingJobName=job_name)
#             print(f"Stopped processing job: {job_name}")
#         except sagemaker_client.exceptions.ClientError as e:
#             if 'ValidationException' in str(e):
#                 print(f"Job {job_name} is already stopped or does not exist.")
#             else:
#                 raise

#         try:
#             sagemaker_client.delete_processing_job(ProcessingJobName=job_name)
#             print(f"Deleted processing job: {job_name}")
#         except sagemaker_client.exceptions.ClientError as e:
#             print(f"Error deleting job {job_name}: {e}")

# delete_processing_jobs(processing_job_names)
