==> ml/notebooks/03-sagemaker-containers.ipynb <==
# imports

import time
import boto3
import botocore
import numpy as np  
import pandas as pd  
import sagemaker
from time import gmtime, strftime, sleep
from sagemaker.sklearn.processing import SKLearnProcessor
from sagemaker.processing import ProcessingInput, ProcessingOutput
from sklearn.metrics import roc_auc_score
from sagemaker.experiments.run import Run, load_run

sagemaker.__version__

boto_session = boto3.Session()
region = boto_session.region_name
print(f"Region--{region}")
sm_role = sagemaker.get_execution_role()
print(f"Sagemaker Role--{sm_role}")
sagemaker_session = sagemaker.Session()
default_bucket = sagemaker_session.default_bucket()
print(f"Default bucket--{default_bucket}")

# SAGEMAKER CONTAINERS


CREATE SageMaker processing and training jobs
In this step you move data processing and model training into SageMaker Docker containers and use SageMaker Python SDK to interact with SageMaker.



* **SageMaker** makes use of Docker containers to enable developers to process data, train and deploy models. 

* **SageMaker** also provides pre-build containers with popular data processing frameworks and ML algorithms. All SageMaker built-in algorithms are delivered as Docker containers.

* **Containers** allow developers and data scientists to package software into standardized units that run consistently on any platform that supports Docker. 

* **Containers** ensure that code, runtime, system tools, system libraries, and settings are all in the same place, isolating them from the execution environment. It guarantees a consistent runtime experience regardless of where a container is being run.




# process data in sagemaker processing jobs

Process data in SageMaker processing job

* Use **SageMaker Processing** by simply providing a Python data preprocessing script and choosing a SageMaker SDK processor class. 
You must upload the input data to S3 and specify an S3 location for output data. 

* **SageMaker Processing** automatically loads the input data from S3 and uploads transformed data back to S3 when the job is complete. 

* The processing container image can either be an Amazon SageMaker built-in image or a custom image that you provide. 

* The underlying infrastructure for a Processing job is fully managed by Amazon SageMaker. 

* Cluster resources are provisioned for the duration of your job, and cleaned up when a job completes.

* Your input data must be stored in an Amazon S3 bucket. Alternatively, you can use Amazon Athena or Amazon Redshift as input sources.



# # import config
# import os
# import yaml
# with open(os.path.abspath(os.path.join(os.pardir,"config.yml")),"r") as f:
#     config = yaml.load(f,Loader=yaml.FullLoader)
# print(config)

# !aws s3 ls {config['s3-bucket']['bucket_name']}/{config['s3-bucket']['bucket_prefix']} --recursive

# write preprocessing script 
> for pre-processing done on the data set

%%writefile ../scripts/preprocessing.py

import pandas as pd
import numpy as np
import argparse
import os

# Treating 'yr' column as Ordinal categorical variable, assign higher value to 2012
yr_mapping = {2011: 0, 2012: 1}
# Treat 'mnth' column as Ordinal categorical variable, and assign values accordingly
mnth_mapping = {'January': 0, 'February': 1, 'December': 2, 'March': 3, 'November': 4, 'April': 5,
                'October': 6, 'May': 7, 'September': 8, 'June': 9, 'July': 10, 'August': 11}
# Treat 'season' column as Ordinal categorical variable, and assign values accordingly
season_mapping = {'spring': 0, 'winter': 1, 'summer': 2, 'fall': 3}
# Map weather situation
weather_mapping = {'Heavy Rain': 0, 'Light Rain': 1, 'Mist': 2, 'Clear': 3}
# Map holiday
holiday_mapping = {'Yes': 0, 'No': 1}
# Map workingday
workingday_mapping = {'No': 0, 'Yes': 1}
# Map hour
hour_mapping = {'4am': 0, '3am': 1, '5am': 2, '2am': 3, '1am': 4, '12am': 5, '6am': 6, '11pm': 7, '10pm': 8,
                '10am': 9, '9pm': 10, '11am': 11, '7am': 12, '9am': 13, '8pm': 14, '2pm': 15, '1pm': 16,
                '12pm': 17, '3pm': 18, '4pm': 19, '7pm': 20, '8am': 21, '6pm': 22, '5pm': 23}


# find numerical and categorical variables
def get_feature_categories(df):
    unused_colms = ['dteday', 'casual', 'registered']   # unused columns will be removed at later stage
    target_col = ['cnt']

    numerical_features = []
    categorical_features = []

    for col in df.columns:
        if col not in target_col + unused_colms:
            if df[col].dtypes == 'float64':
                numerical_features.append(col)
            else:
                categorical_features.append(col)
    print('Number of numerical variables: {}'.format(len(numerical_features)),":" , numerical_features)
    print('Number of categorical variables: {}'.format(len(categorical_features)),":" , categorical_features)
    return numerical_features , categorical_features , unused_colms


 # Working on `dteday` column to extract year and month
def get_year_and_month(dataframe):
    df = dataframe.copy()
    # convert 'dteday' column to Datetime datatype
    df['dteday'] = pd.to_datetime(df['dteday'], format='%Y-%m-%d')
    # Add new features 'yr' and 'mnth
    df['yr'] = df['dteday'].dt.year
    df['mnth'] = df['dteday'].dt.month_name()
    print("Added yr and month variables to the dataframe--")
    return df


# Function to impute weekday by extracting day name from the date column

def impute_weekday(dataframe):
    df = dataframe.copy()
    wkday_null_idx = df[df['weekday'].isnull() == True].index
    # print(len(wkday_null_idx))
    df.loc[wkday_null_idx, 'weekday'] = df.loc[wkday_null_idx, 'dteday'].dt.day_name().apply(lambda x: x[:3])
    print("imputed weekday in dataframe--")
    return df


# Function to handle outliers for a single column

def handle_outliers(dataframe, colm):
    df = dataframe.copy()
    q1 = df.describe()[colm].loc['25%']
    q3 = df.describe()[colm].loc['75%']
    iqr = q3 - q1
    lower_bound = q1 - (1.5 * iqr)
    upper_bound = q3 + (1.5 * iqr)
    for i in df.index:
        if df.loc[i,colm] > upper_bound:
            df.loc[i,colm]= upper_bound
        if df.loc[i,colm] < lower_bound:
            df.loc[i,colm]= lower_bound
    print(f"outliers handled in column--{colm}")
    return df


# Treating 'weekday' column as a Nominal categorical variable, perform one-hot encoding
from sklearn.preprocessing import  OneHotEncoder
# build encoder
encoder = OneHotEncoder()

## Function for pre-processing the dataset

def pre_process(dataframe):
    df = dataframe.copy()
    df = get_year_and_month(df)
    df = impute_weekday(df)
    df['weathersit'].fillna('Clear', inplace=True)
    numerical_features , categorical_features , unused_colms = get_feature_categories(df)
    for col in numerical_features:
        df = handle_outliers(df, col)
    df['yr'] = df['yr'].apply(lambda x: yr_mapping[x])
    df['mnth'] = df['mnth'].apply(lambda x: mnth_mapping[x])
    df['season'] = df['season'].apply(lambda x: season_mapping[x])
    df['weathersit'] = df['weathersit'].apply(lambda x: weather_mapping[x])
    df['holiday'] = df['holiday'].apply(lambda x: holiday_mapping[x])
    df['workingday'] = df['workingday'].apply(lambda x: workingday_mapping[x])
    df['hr'] = df['hr'].apply(lambda x: hour_mapping[x])
    encoder.fit(df[['weekday']])
    try:
        enc_wkday_features = encoder.get_feature_names_out(['weekday'])
    except Exception as e:
        print(f"Exception occured on get_features_names_out--{e}")
        enc_wkday_features = [f'weekday_{i}' for i in range(encoder.categories_[0].shape[0])]
    encoded_weekday = encoder.transform(df[['weekday']]).toarray()
    df[enc_wkday_features] = encoded_weekday
    # drop not required columns
    unused_colms.append('weekday')
    df.drop(labels = unused_colms, axis = 1, inplace = True)
    print("preprocessing done !")
    return df




def _parse_args():
    
    parser = argparse.ArgumentParser()
    # Data, model, and output directories
    # model_dir is always passed in from SageMaker. By default this is a S3 path under the default bucket.
    parser.add_argument('--filepath', type=str, default='/opt/ml/processing/input/')
    parser.add_argument('--filename', type=str, default='bike-sharing-dataset.csv')
    parser.add_argument('--outputpath', type=str, default='/opt/ml/processing/output/')
    
    return parser.parse_known_args()


if __name__=="__main__":
    

    # Process arguments
    args, _ = _parse_args()
    
    # Verify if the file exists
    if not os.path.exists(os.path.join(args.filepath, args.filename)):
        raise FileNotFoundError(f"File not found: {input_file_path}")

    
    target_col = "cnt"
    
    # Load data
    df_data = pd.read_csv(os.path.join(args.filepath, args.filename))
    try:
        df_data.drop(['Unnamed: 0'],axis=1 , inplace=True)
    except Exception as e:
        print(f"Exception occured in dropping column--unnamed: 0-->{e}")
    processed_data = pre_process(df_data)

    # Shuffle and splitting dataset
    train_data, validation_data, test_data = np.split(
    processed_data.sample(frac=1, random_state=1729),
    [int(0.7 * len(processed_data)), int(0.9 * len(processed_data))],)

    print(f"Data split > train:{train_data.shape} | validation:{validation_data.shape} | test:{test_data.shape}")

    
    # Save datasets locally
    print(f"output path--{args.outputpath}")
    train_data.to_csv(os.path.join(args.outputpath, 'train/train.csv'), index=False, header=True)
    validation_data.to_csv(os.path.join(args.outputpath, 'validation/validation.csv'), index=False, header=True)
    test_data[target_col].to_csv(os.path.join(args.outputpath, 'test/test_y.csv'), index=False, header=True)
    test_data.drop([target_col], axis=1).to_csv(os.path.join(args.outputpath, 'test/test_x.csv'), index=False, header=True)
    print(f"output path--{args.outputpath}")
    # Save the baseline dataset for model monitoring
    processed_data.drop([target_col], axis=1).to_csv(os.path.join(args.outputpath, 'baseline/baseline.csv'), index=True, header=False)
    
    print("## Processing complete. Exiting.")


# set s3 paths

def get_default_bucket():
    sagemaker_session = sagemaker.Session()
    default_bucket = sagemaker_session.default_bucket()
    print(f"Default bucket--{default_bucket}")
    return default_bucket

PREFIX = 'bits-webminar-june-24/'
DEFAULT_BUCKET = get_default_bucket()

s3_client = boto3.client("s3")

# display a list of s3 object
objects = s3_client.list_objects_v2(Bucket=default_bucket)
for obj in objects.get('Contents', []):
    # print(obj)
    file_key = obj['Key']
    # print(file_key)
    if file_key == "bits-webminar-june-24/":
        continue
    else:
        s3_file_path = f"s3://{default_bucket}/{file_key}"
        print(s3_file_path)


DEFAULT_BUCKET

input_s3_url ='s3://sagemaker-eu-north-1-058264393695/bits-webminar-june-24/bike-sharing-dataset.csv'
print(f"input_s3_url--> {input_s3_url}")
train_s3_url ='s3://sagemaker-eu-north-1-058264393695/bits-webminar-june-24/container-jobs/train'
print(f"train_s3_url--> {train_s3_url}")
test_s3_url ='s3://sagemaker-eu-north-1-058264393695/bits-webminar-june-24/container-jobs/test'
print(f"test_s3_url--> {test_s3_url}")
validation_s3_url ='s3://sagemaker-eu-north-1-058264393695/bits-webminar-june-24/container-jobs/validation'
print(f"test_s3_url--> {test_s3_url}")
baseline_s3_url = f"s3://{DEFAULT_BUCKET}/{PREFIX}container-jobs/baseline"
print(f"baseline_s3_url--> {baseline_s3_url}")

# framework_version = "0.23-1"
framework_version = "1.0-1"
processing_instance_type = "ml.t3.medium"
processing_instance_count = 1

# Create an experiment run

# experiment_name = "bits-webminar-june-24-21-06-2024-15-11-46"
experiment_name = "bike-ride-sharing"

# using the existing experiment 

import sys
from pathlib import Path

root_dir = Path('../scripts').resolve().parents[0] 
print(root_dir)
sys.path.append(str(root_dir))

import scripts.utils as utils
run_suffix = utils.get_current_ist().strftime("%d-%m-%Y-%H-%M-%S")
run_suffix

sagemaker_session = sagemaker.Session()
sm_client = sagemaker_session.sagemaker_client

# Create an experiment config

run_name = f"container-processing"

with Run(experiment_name=experiment_name,
         run_name=run_name,
         run_display_name="container-processing",
         sagemaker_session=sagemaker_session
        ) as run:
    run.log_parameters(
        {
            "train": 0.7,
            "validate": 0.2,
            "test": 0.1
        }
    )
   
    experiment_config = run.experiment_config
    # time.sleep(8) # wait until resource tags are propagated to the run

# Create a sagemaker sklearn processor --IMAGE

sm_role = sagemaker.get_execution_role()
print(f"Sagemaker Role--{sm_role}")

sklearn_processor = SKLearnProcessor(
    framework_version=framework_version,
    role=sm_role,
    instance_type=processing_instance_type,
    instance_count=processing_instance_count, 
    base_job_name='container-processing',
    sagemaker_session=sagemaker_session,
)

# Define processing job

processing_inputs = [
        ProcessingInput(
            source=input_s3_url, 
            destination="/opt/ml/processing/input",
            s3_input_mode="File",
            # s3_data_distribution_type="ShardedByS3Key",
            s3_data_distribution_type="FullyReplicated",
        )
    ]

processing_outputs = [
        ProcessingOutput(
            output_name="train_data", 
            source="/opt/ml/processing/output/train",
            destination=train_s3_url,
        ),
        ProcessingOutput(
            output_name="validation_data", 
            source="/opt/ml/processing/output/validation", 
            destination=validation_s3_url
        ),
        ProcessingOutput(
            output_name="test_data", 
            source="/opt/ml/processing/output/test", 
            destination=test_s3_url
        ),
        ProcessingOutput(
            output_name="baseline_data", 
            source="/opt/ml/processing/output/baseline", 
            destination=baseline_s3_url
        ),
    ]

# Start the sagemaker processing job

try:
    sklearn_processor.run(
        inputs=processing_inputs,
        outputs=processing_outputs,
        code='../scripts/preprocessing.py',
        wait=True,
        experiment_config=experiment_config,
        # arguments = ['arg1', 'arg2'],
    )
except botocore.exceptions.ClientError as e:
    if e.response['Error']['Code'] == 'AccessDeniedException':
        print(f"Ignore AccessDeniedException: {e.response['Error']['Message']} because of the slow resource tag auto propagation")
    else:
        raise e

# Model training in SageMaker training job

# get training container uri
# region = config['sagemaker']['region']
training_image = sagemaker.image_uris.retrieve("xgboost", region=region, version="1.5-1")
print(training_image)

train_instance_count = 1
train_instance_type = "ml.m5.xlarge"

# Define where the training job stores the model artifact
output_s3_url = "s3://sagemaker-eu-north-1-058264393695/bits-webminar-june-24/container-jobs/training-output"
output_s3_url

# sagemaker-estimator

# Instantiate an XGBoost estimator object
estimator = sagemaker.estimator.Estimator(
    image_uri=training_image,  # XGBoost algorithm container
    instance_type=train_instance_type,  
    instance_count=train_instance_count,  
    role=sm_role,  
    max_run=20 * 60,  # Maximum allowed active runtime
    output_path=output_s3_url, # S3 location 
    sagemaker_session=sagemaker_session, # Session object  manages interactions with SageMaker API and AWS services
    base_job_name="xgboost-training", # Prefix for training job name
)

# define its hyperparameters
estimator.set_hyperparameters(
    num_round=150, # the number of rounds to run the training
    max_depth=3, # maximum depth of a tree
    eta=0.5, # step size shrinkage used in updates to prevent overfitting
    alpha=2.5, # L1 regularization term on weights
    objective="reg:squarederror",
    eval_metric="rmse", # evaluation metrics for validation data
    subsample=0.8, # subsample ratio of the training instance
    colsample_bytree=0.8, # subsample ratio of columns when constructing each tree
    min_child_weight=3, # minimum sum of instance weight (hessian) needed in a child
    early_stopping_rounds=10, # the model trains until the validation score stops improving
    verbosity=1, # verbosity of printing messages
)

s3_input_train = sagemaker.inputs.TrainingInput(train_s3_url, content_type='csv')
s3_input_validation = sagemaker.inputs.TrainingInput(validation_s3_url, content_type='csv')

training_inputs = {'train': s3_input_train, 'validation': s3_input_validation}

try:
    run_suffix = utils.get_current_ist().strftime("%d-%m-%Y-%H-%M-%S")
    print(f"run_suffix--{run_suffix}")
    run_name = f"container-training-{run_suffix}"

    with Run(experiment_name=experiment_name,
             run_name=run_name,
             run_display_name="container-training",
             sagemaker_session=sagemaker_session
            ) as run:
        
        estimator.fit(
            training_inputs,
            wait=True,
            logs=False,
        ) 
except botocore.exceptions.ClientError as e:
    if e.response['Error']['Code'] == 'AccessDeniedException':
        print(f"Ignore AccessDeniedException: {e.response['Error']['Message']} because of the slow resource tag auto propagation")
    else:
        raise e

# Output model performance

if estimator._current_job_name:
    training_job_name = estimator._current_job_name

training_job_name

metrics = None
while not metrics:
    metrics = sm_client.describe_training_job(
        TrainingJobName=training_job_name
        ).get("FinalMetricDataList")

    if not metrics:
        print(f"Training job {training_job_name} hasn't finished yet!")
        time.sleep(10)

train_rmse = float([m['Value'] for m in metrics if m['MetricName'] == 'train:rmse'][0])
validate_rmse = float([m['Value'] for m in metrics if m['MetricName'] == 'validation:rmse'][0])

print(f"Train RMSE: {train_rmse:.2f}, Validate RMSE: {validate_rmse:.2f}")



# Print the S3 path to the model artifact:
estimator.model_data

destination_s3_uri = "s3://sagemaker-eu-north-1-058264393695/bits-webminar-june-24/container-jobs/data-capture"
destination_s3_uri

# Validate model -- deploy endpoint

# Real-time endpoint
from time import gmtime
endpoint_name = f"bits-webminar-xgboost-endpoint-{strftime('%d-%H-%M-%S', gmtime())}"

try:
    predictor = estimator.deploy(
        initial_instance_count=1,
        instance_type="ml.m5.xlarge",
        # instance_type="ml.t3.medium",
        
        wait=False,  
        # data capture here, for experiment with monitoring:
        data_capture_config=sagemaker.model_monitor.DataCaptureConfig(
            enable_capture=True,
            sampling_percentage=100,
            destination_s3_uri = "s3://sagemaker-eu-north-1-058264393695/bits-webminar-june-24/container-jobs/data-capture"
        ),
        endpoint_name=endpoint_name,
        serializer=sagemaker.serializers.CSVSerializer(),
        deserializer=sagemaker.deserializers.CSVDeserializer(),
    )
except botocore.exceptions.ClientError as e:
    if e.response['Error']['Code'] == 'AccessDeniedException':
        print(f"Ignore AccessDeniedException: {e.response['Error']['Message']} because of the slow resource tag auto propagation")
        predictor = sagemaker.predictor.Predictor(endpoint_name=endpoint_name,
                                                  sagemaker_session=sagemaker_session,
                                                  serializer=sagemaker.serializers.CSVSerializer(),
                                                  deserializer=sagemaker.deserializers.CSVDeserializer(),
                                                 )
    else:
        raise e

# Wait until the endpoint has the status InService
waiter = sagemaker_session.sagemaker_client.get_waiter('endpoint_in_service')
waiter.wait(EndpointName=endpoint_name)

# !aws s3 cp $test_s3_url/test_x.csv /home/sagemaker-user/siemens-poc/test_x.csv
# !aws s3 cp $test_s3_url/test_y.csv /home/sagemaker-user/siemens-poc/test_y.csv

test_s3_url

# Batch processing

default_bucket


# s3://sagemaker-eu-north-1-058264393695/bits-webminar-june-24/container-jobs/test/test_x.csv

response = s3_client.list_objects_v2(Bucket=default_bucket)

if 'Contents' in response:
    for obj in response['Contents']:
        print(obj['Key'])
else:
    print("No objects found in the bucket.")


# test_x = pd.read_csv("tmp/test_x.csv", names=[f'{i}' for i in range(59)])
# test_y = pd.read_csv("tmp/test_y.csv", names=['y'])

import io
s3_client = boto3.client("s3")
file_key = "bits-webminar-june-24/container-jobs/test/test_x.csv"
response = s3_client.get_object(Bucket=default_bucket, Key=file_key)
file_content = response['Body'].read()
test_x = pd.read_csv(io.BytesIO(file_content), encoding='utf-8')
test_x.head(3)

file_key  = "bits-webminar-june-24/container-jobs/test/test_y.csv"
response = s3_client.get_object(Bucket=default_bucket, Key=file_key)
file_content = response['Body'].read()
test_y = pd.read_csv(io.BytesIO(file_content), encoding='utf-8')
test_y.head(3)

## sagemaker-predictor

endpoint_name = "bits-webminar-xgboost-endpoint-27-09-16-13"
predictor = sagemaker.predictor.Predictor(endpoint_name=endpoint_name,
                                          sagemaker_session=sagemaker_session,
                                          serializer=sagemaker.serializers.CSVSerializer(),
                                          deserializer=sagemaker.deserializers.CSVDeserializer(),
                                         )

predictions = np.array(predictor.predict(test_x.values), dtype=float).squeeze()
predictions

test_x.values[0]

# evaluate predictions 
test_results = pd.concat(
    [
        pd.Series(predictions, name="y_pred", index=test_x.index),
        test_x,
    ],
    axis=1,
)
test_results.head()

# test_results.iloc[0].values

# pd.crosstab(
#     index=test_y['cnt'].values,
#     columns=np.round(predictions), 
#     rownames=['actuals'], 
#     colnames=['predictions']
# )

from sklearn.metrics import mean_squared_error
test_rmse = mean_squared_error(test_y, test_results["y_pred"])
print(f"Test-rmse: {test_rmse:.2f}")

# Stream and Load the Model Directly from S3

import boto3
import io
import tarfile
import tempfile
import xgboost as xgb

# model_file_key = "bits-webminar-june-24/container-jobs/training-output/xgboost-training-2024-06-23-03-37-48-851/output/model.tar.gz"
model_file_key = "bits-webminar-june-24/container-jobs/training-output/xgboost-training-2024-06-27-09-04-19-728/output/model.tar.gz"

s3 = boto3.resource('s3')

model_file_stream = io.BytesIO()
s3.Bucket(default_bucket).download_fileobj(model_file_key, model_file_stream)

# Set the stream position to the start
model_file_stream.seek(0)

# Extract the model file from the tar.gz archive in memory
with tarfile.open(fileobj=model_file_stream, mode="r:gz") as tar:
    # Extract the model file to a temporary file
    with tempfile.NamedTemporaryFile(delete=False) as tmp_file:
        for member in tar.getmembers():
            if member.name == 'xgboost-model':  # Adjust the file name as needed
                tmp_file.write(tar.extractfile(member).read())
                tmp_file.flush()
                break

# Load the XGBoost model from the temporary file
model = xgb.Booster()
model.load_model(tmp_file.name)



import io
file_key = "bits-webminar-june-24/container-jobs/test/test_x.csv"
s3_client = boto3.client("s3")
response = s3_client.get_object(Bucket=default_bucket, Key=file_key)
file_content = response['Body'].read()
X_test = pd.read_csv(io.BytesIO(file_content), encoding='utf-8')
# print(X_test.head(3))
test_x = xgb.DMatrix(X_test.values)

file_key = "bits-webminar-june-24/container-jobs/test/test_y.csv"
response = s3_client.get_object(Bucket=default_bucket, Key=file_key)
file_content = response['Body'].read()
y_test = pd.read_csv(io.BytesIO(file_content), encoding='utf-8')
# y_test.head(3)
test_y = y_test.to_numpy()

# Read test data
# print(test_x)
# print(test_y)


# Run predictions
predictions = np.array(model.predict(test_x), dtype=float).squeeze()
print(predictions)


# Evaluate predictions
test_results = pd.concat([pd.Series(predictions, name="y_pred", index=X_test.index),X_test,],axis=1,)
test_results.head()


from sklearn.metrics import mean_squared_error
test_rmse = mean_squared_error(test_y, test_results["y_pred"])
report_dict = {"regression_metric":{"test_rmse":{"value":test_rmse}}}
print(f"Test-rmse: {test_rmse:.2f}")




report_dict
# test_results
default_bucket

def upload_file_to_s3(df,file_key):
    csv_buffer = io.StringIO()
    df.to_csv(csv_buffer)
    BUCKET_NAME  = default_bucket
    PREFIX = 'bits-webminar-june-24/'
    KEY = f"{PREFIX}{file_key}"
    try:
        s3_client.put_object(
            Body=csv_buffer.getvalue(), 
            Bucket = BUCKET_NAME, 
            Key = KEY,
        )
        print("Upload Successful")
    except FileNotFoundError:  
        print("The file was not found")
    except botocore.exceptions.ClientError as error:
        if error.response['Error']['Code'] == "404":
            print("The object does not exist")
        else:
            raise

upload_file_to_s3(test_results,'prediction_baseline/prediction_baseline.csv')

# Save evaluation report
# import pathlib
# import json
# pathlib.Path(output_dir).mkdir(parents=True, exist_ok=True)
# with open(f"{output_dir}/evaluation.json", "w") as f:
#     f.write(json.dumps(report_dict))



# Save prediction baseline file - we need it later for the model quality monitoring
# test_results.to_csv(os.path.join(output_prediction_path, 'prediction_baseline/prediction_baseline.csv'), index=False, header=True)

# delete processing jobs 

import boto3
sagemaker_client = boto3.client('sagemaker')
response = sagemaker_client.list_processing_jobs()
for job in response['ProcessingJobSummaries']:
    print(job['ProcessingJobName'])
